Была произведена отладка сессии для нескольких клиентов

+++ sources/base/CanFTP_BaseTypes_Defines.h
@@ -17,17 +17,17 @@ typedef uint8_t CanFTP_Logical_t;
 // Тип серийного номера устройства
 typedef uint32_t CanFTP_DeviceSerial_t;
 // Тип идентификатора устройства
-typedef uint8_t CanFTP_DeviceIdentifier_t;
+typedef uint16_t CanFTP_DeviceIdentifier_t;
 // Тип идентификатора устройства
 typedef uint16_t CanFTP_DeviceType_t;
 // Тип идентификатора устройства
 typedef uint8_t CanFTP_DeviceSoftwareVersionPart_t;
 // Тип кода сессии сессии
-typedef uint8_t CanFTP_SessionCode_t;
+typedef uint16_t CanFTP_SessionCode_t;
 // Тип индекса фрейма в рамках блока кода
-typedef uint8_t CanFTP_FrameIndex_t;
+typedef uint16_t CanFTP_FrameIndex_t;
 // Тип кода устройства
-typedef uint8_t CanFTP_DeviceCode_t;
+typedef uint16_t CanFTP_DeviceCode_t;
 // Тип кода устройства
 typedef uint16_t CanFTP_DeviceSerialCode_t;
 // Тип мндекса блока кода
@@ -37,13 +37,13 @@ typedef uint16_t CanFTP_BlockLength_t;
 // Тип версии протокола
 typedef uint16_t CanFTP_ProtocolVersion_t;
 // Тип индекса страницы записи
-typedef uint8_t CanFTP_PageIndex_t;
+typedef uint16_t CanFTP_PageIndex_t;
 // Тип размера передаваемого файла
 typedef uint32_t CanFTP_FileLength_t;
 // Тип параметра количества повтореных отправок
-typedef uint8_t CanFTP_SendingRepeate_t;
+typedef uint16_t CanFTP_SendingRepeate_t;
 // Тип параметра периодов повторных отправок
-typedef uint8_t CanFTP_SendingRepeateInterval_t;
+typedef uint16_t CanFTP_SendingRepeateInterval_t;
 // Тип кодификатора идентификаторов сообщений
 typedef uint32_t CanFTP_CanMessageCodificator_t;
